Version 0.0.3.143:
- Einige Memory-Leaks im Videomanager und VBE beseitigt
- Kleinere Optimierungen im Paging-Modul
- TransferFlag bei UHCI, OHCI und EHCI eliminiert (redundant durch andere Flags, inkonsequente Nutzung)
- UHCI kompiliert nun auch ohne _UHCI_DIAGNOSIS_

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@1344 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/storage/uhci.h
@@ -194,7 +194,7 @@ typedef struct uhci
     mutex_t*       qhLock;                   // mutex for access on the QH
     bool           enabledPortFlag;          // root ports enabled
     uhci_port_t*   ports[UHCIPORTMAX];       // root ports
-    uint32_t       connected[UHCIPORTMAX];   // port connected (counter)
+    uint8_t        connected[UHCIPORTMAX];   // port connected (counter)
     bool           run;                      // hc running (RS bit)
     uint8_t        num;                      // Number of the UHCI
 } uhci_t;
